Remove landscape/portrait conditional style sheets
Issue #23:

The pop-up keyboard on a phone in portrait mode was causing the viewport
to trigger the landscape layout.

The landscape layout happened to define a min-width that's wider than
the phone screen's portrait width.

There is no desired layout difference at the moment between portrait and
landscape viewports on mobile, so the solution for now is to get rid of
these style sheets altogether.