Implement concept of "allocations" in table view
Issue #28:

This affects the table view (visible in wide viewports)

- New column "Paid by" that displays the name of the payer
- Column "Amount" now displays the nominal amount of the record
- Original column "Amount" renamed to "Effective" because it displays
the amount that's effective towards the total cost of the event

Additional view for the "Participants" section: expense allocations, to
complement the existing "contributions" view.

The allocation values are present, but are currently hidden.
A toggle control will be implemented in the near future to toggle
between the allocation and the contribution views.