Make variance algorithm aware of allocations
Issue #30:
The variance-calculating algorithm used to assume that all expenses were
attributed to the entire group, and simply divided the event total by
the number of participants.
This commit enhances the algorithm to be aware of expense allocations to
varying sub-groups.
Each participant's variance is calculated as the difference of the
participant's contribution and the debt allocated to the participant.