Add allocations selector to the expense record form
Issue #32:
Up to now, it was assumed that all expenses apply to the entire group.
This commit adds a new field to the expense form: allocations (labeled
"Covers", as in, the expense covers participants X, Y, Z)
The user can toggle between two modes: "entire group" and "select
participants"
The second mode brings up a list of check-boxes labeled with the
participant names; The user can select any non-empty combination of
participants.
Combinations of size 1 are allowed, which denotes personal expenses.