Introduce "default" event concept
Issue #37:
This adds a new column to the participant table: "is_default"
When a participant record is flagged as "default", that event's page
becomes the user's landing page upon login.
If a user has no default event designated, the landing page is their
account page.
Controls have been added to the account page to select the default
event.